requirements.txt
django-twitter-bootstrap